getting the last version of the main

+++ myproject/myapp/models.py
@@ -8,13 +8,12 @@ from asgiref.sync import async_to_sync
 from channels.layers import get_channel_layer
 
 class User(AbstractUser):
-    # other fields
     image = models.URLField(max_length=255, null=True, blank=True)
     is_online = models.BooleanField(default=False)
     is_2fa_enabled = models.BooleanField(default=False)
     wins = models.IntegerField(default=0)
     losses = models.IntegerField(default=0)
-    level = models.IntegerField(default=0)
+    level = models.FloatField(default=0)
     winrate = models.FloatField(default=0)
     rank = models.IntegerField(default=0)
     total_goals_scored = models.IntegerField(default=0)
@@ -40,7 +39,6 @@ class User(AbstractUser):
         return self.username
 
 class Achievement(models.Model):
-    # user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     achievement = models.CharField(max_length=255)
     date = models.DateTimeField(auto_now_add=True)
     description = models.TextField()
@@ -108,7 +106,3 @@ class Notification(models.Model):
 
     def __str__(self):
         return f"{self.notification_type} for {self.recipient.username}"
-
-# after making a new model or making changes to a models we use:
-# python manage.py makemigrations
-# python manage.py migrate
